issue-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
keybert>=0.7.0
konlpy>=0.6.0
pdfminer.six>=20221105
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
python-multipart>=0.0.5
